refactor(webapp): move the palette scorer to lib/fuzzy.ts
Pure move, no behavior change, so the commit that follows is only the
behavior. Follows the repo convention that pure logic leaves the
component (lib/heat.ts, lib/csv.ts, lib/scroll.ts).
