+++ internal/webapp/frontend/src/lib/fuzzy.ts
@@ -1,0 +1,40 @@
+/* The ⌘K palette's matcher. Lives here, not in the component, so it can be
+   unit-tested without a DOM (see fuzzy.test.ts). Palette.tsx keeps only
+   rendering. */
+
+/* Subsequence fuzzy match. Returns a score (higher = better), plus the
+   matched positions for highlighting; null when it doesn't match. */
+export function fuzzy(query: string, text: string): { score: number; hits: number[] } | null {
+  if (!query) return { score: 0, hits: [] };
+  const q = query.toLowerCase();
+  const t = text.toLowerCase();
+  let ti = 0,
+    score = 0,
+    streak = 0;
+  const hits: number[] = [];
+  for (let qi = 0; qi < q.length; qi++) {
+    const found = t.indexOf(q[qi], ti);
+    if (found === -1) return null;
+    streak = found === ti ? streak + 1 : 1;
+    score += streak * 3; // consecutive runs
+    if (found === 0 || "/ -_.".includes(t[found - 1])) score += 8; // word starts
+    hits.push(found);
+    ti = found + 1;
+  }
+  score -= Math.floor(t.length / 8); // mild preference for short targets
+  return { score, hits };
+}
+
+/* Match a query against a label, tolerating a simple English plural so
+   "ideas" still finds idea.md. Tries the raw query first, then a lightly
+   de-pluralized form (…ies→…y, …es→…, …s→…). */
+export function fuzzyStemmed(query: string, label: string) {
+  const m = fuzzy(query, label);
+  if (m) return m;
+  const q = query.toLowerCase();
+  let stem: string | null = null;
+  if (q.length > 3 && q.endsWith("ies")) stem = q.slice(0, -3) + "y";
+  else if (q.length > 3 && q.endsWith("es")) stem = q.slice(0, -2);
+  else if (q.length > 2 && q.endsWith("s")) stem = q.slice(0, -1);
+  return stem ? fuzzy(stem, label) : null;
+}
